MSTest based functional tests for service fabric provider.
(User SF remoting as communication mechanism for tests)

However deploying the local cluster is still a manul pre-req step for tests to run (noted as a todo).

Also, at this point, the next step would be to put all existing functional tests for service bus provider
into a separate dll and reuse them to test the service fabric provider to understand the gaps.