Implementing one to-do which I think we will need : Support purging state for old orchestrations.
There were some tradeoffs made such as not being able to query for state of orchestration that completed an hour ago,
for our purposes where we have short running workflows, we are ok (because we also want the instance store to be
locally kept but not necessarily use Azure storage).

We can perhaps later implement integration with azure instance store from this class eventually.