Changes for trying out timeout wrapper in orchestrations.
Note that timer needs to be explicitly cancelled once the main task finished because otherwise the orchestration will wait for timer to complete firing taking it more time than needed.

This is by design in Durable task framework. (It would be same for main task in the sense that if timer fires first, somehow main task needs to be cancelled, for the test implementation
I considered that optional and left it out)