A test tooling app to inspect stuck orchestrations - this is just a console program for the test app and uses the same API that was introduced for tooling a few days ago.
Implemented something quick/simple with good enough experience/flow.

It could be evolved in future based on what is most needed and valuable.

(for example using a command pattern to implement individual commands, or to use a proper console option parser to be more automated or use interconnected flow
like using the output of list orchestrations and being able to point to one value for further actions like terminate / get runtime state etc)