Fixing an issue identified in failover testing where some orchestrations are stuck.
It was a failrly consistent repro and the runtime session state of all stuck orchestration is null and the root
cause is that Complete must have updated the session state to null and Release was not called
because of node failure and so the orchestration was never completed and instance store never updated.

The fix is a simple and relatively safer fix to not update session state with null when orchestration is completed.
Perhaps another fix would be to keep Release simple and let Complete drop the session instead
when orchestration is complete - but that's slightly riskier at this point.

Will check this in first and will try the other fix later.

Minor change to chaos runner test program to induce failures more frequently.