Revert "Correct way to fix stuck orchestrations issue (this was a slight riskeier fix than previous one)"
This reverts commit 272db89f61dc76f2dd920c26fbf15d7a077dbba5.

There's an issue with dropping session as part of complete - dropping a
session involves 2 steps : Dropping session messages dictionary, and
removing the row for session in sessions dictionary. Dropping session
messages dictionary is done in a background thread and hence not part of
transaction. So if the transaction failed after retries, we will end up
not updating the completed session state but would have lost the final
messages in the orchestration that make it to complete stage.

We could do dropping session messages absolutely as last part of the
transaction (so usually the previous steps fail while taking lock with
Timeouts and we never initiate the last step) but it's just safer to
update session state and complete messages in this transaction and do Drop
Session (both the session row and session messages dictionary in a
subsequent transaction) - that way we atleast would have updated the
session state with the completed messages.

Also, we want to update the instance store with the complted status only
when we remove the session row in sessions dictionary so that creating an
orchestration with the same instance id once it's completed works as
expected. So the right fix would be the previous fix where as part of
Release method (really just a subsequent transaction) we update instance
store and drop the session. In this case, even if this transaction is
failed but we ended up dropping session messages, we were anyway deleting
an empty session messages dictionary and we captured the latest runtime
state with the complted state so the next accept will have chance to
complete the session.