I missed one place where the transaction to update the state to completed state succeeded, but the transaction to complete the orchestration failed and the session was abandoned and tried again.
In this case, the code before called ReleaseTaskOrchestrationWorkItemAsync which used
to handle the completed orchestration but not anymore. Replacing that with
the new method introduced for handling completed orchestration.

Also, when that also fails, we want to make sure we abandon the session so that
it will be picked up again and eventually remove the orchestration entries.

Perhaps I need to re-think this scenario again, maybe I need to drop the session
messages dictionary in the susequent transaction and remove the dictionary row in the original
transaction.