Fixing a subtle bug with TryEnqueueSession in the way it uses 'AddOrUpdate' from ConcurrentDictionary. from concurrent dictionary.
(AddOrUpdate is not atomic because the update value factory is executed outside the lock in concurrent dictionary,
I have seen with a minimized repro that this is somehow causing a session to be skipped from getting enqueued
in the fetch queue - though I know that it's becase of my use of a local variable
that gets set inside the update value factory and used to decide whether to add to
fetch queue or not, I am still not 100% sure how might that happen).

But with my minimized repro program, I have seen that this fix makes it not repro. (Ran it a bunch of times to make sure).

I am expecting this is the same bug that is causing a few stuck orchestrations in messaging stress.