After reviewing the service bus client functionality for WaitForOrchestation client API which ignores the input executionId and instead wait for latest execution id of given orchestration, I changed the fabric provider client's implementation to match that.
The fabric provider still does not support querying for orchestration state across all executions.
To implement that, we perhaps needs to introduce one more dictionary to track all execution ids for a
given orchestration instance. There are some open questions about this like should we differentiate
the executions from ContinueAsNew to that of different executions caused by user creating the orchestration
with same instanceId after the current one completes. Also it's not clear if this
functionality is desired at all. So until I hear otherwise, I choose to postpone it.