Make large message blob names unique (#326)
We previously generated "unique" blob names using the SequenceNumber
field of the MessageData object. However, these sequence numbers are
generated by the client, This means is no uniqueness guarantee for a given
instance id, as any VM running the application is allowed to send
messages. This meant it was possible for large messages sent from
different VMs to the same instance could overwrite each other.

We now use activity ids, as they are random guids that should
not conflict with other messages, regardless of which client they are
sent from.