Prevent value drift on PendingOrchestrationCount
Currently, when a host lease is lost and reclaimed before its pending messages are processed, a new control queue is spun up, losing track of what messages it has already processed.

In an ideal solution, we would find some way to reclaim that old control queue. This commit tracks a more simple workaround in which we don't rely on the message id being present in the current control queues pending messages in order to decrement PendingOrchestrationCount. This may not fix all of the problems (i.e. duplicate messages may be still be double counted in the lease reaquire scenario). It would still be an improvement over the current scenario however.