Add optional parameter to remove some telemetry for performance reasons
Add a flag to TaskHubWorker and some of its internal classes in order to
flag that we prefer performance over perfectly detailed telemetry.

Currently the only telemetry this field affects is the runtime state
dump. This is a linear operation with the size of the event history,
and even if you filter out trace/information level events, there is
still the computational cost of creating the dump. In efficient
telemetry mode, we only emit the number of events rather than the events
themselves.