Fix edge case with Azure Storage large message implementation
Azure Storage queues can only store 64 KiB. Before, we would put any raw
messages larger than 60 KiB in blob storage and send a reference to the
blob in the queue message. However, a customer has experienced an issue
in which a raw input between 60000 bytes and 61440 bytes created a queue
message larger than 64 KiB, likely due to the metadata we send along
with the input in the queue message.

This PR solves this issue with a two-pronged approach. First, we reduce
the raw input size limit before writing to blob storage from 60 KiB to
60000 bytes. Second, we check the message size after including our
metadata to ensure we are not sending anything greater than 64 KiB, and
if we are, we store the input in blob storage.