Make large message handling more robust
PR #435 attempted to fix large message inconsistencies in how it handled
UTF8 vs UTF16 character counts, but it turns out that the code path that starts new orchestrations relied on the overcounting of characters that used to occur.

This PR fixes that case, and adds more testing with characters of
varying UTF8 and UTF16 sizes to make sure we cover all of the edge
cases.

Addresses #448