Update stale taskhub.json in DurableTask.AzureStorage
The Azure Functions scaling infrastructure relies on the partitionCount
value in the taskhub.json. However, our logic to create this
taskhub.json only updated the blob if it did not exist. This means that
if the taskhub lease store was created with one value of partitionCount,
and the customer updated the partitionCount, the taskhub.json would not
be updated, and the scaling logic would continue to look only at the
original control queues.

This PR aims to update the taskhub.json partition count if it is out of
sync with the provided value in AzureStorageOrchestrationService, so
that customers don't need to update this value by hand to get correct
scaling behavior.