Retry operations on Storage SDK deadlocks instead of app recycle (#500)
The Azure Storage backend will now attempt to retry operations when a
storage deadlock is detected instead of just recycling the app. In order
to do this properly, we had to make sure to cancel the in flight
operation just in case it actually would finish, to avoid performing
duplicate work. We also still need to recycle the app after multiple
deadlock hits to avoid resource leaks.

This PR also properly wraps several missed storage calls in our helper method
so that they have this deadlock mitigation as well.