Reduce storage IO costs for partition management
This PR attempts to address
https://github.com/Azure/azure-functions-durable-extension/issues/1651.
In order to accomplish this, we now only download blob data for
ownership leases when we already own the intent lease.

This change did not play nicely with the old abstractions, which were
largely copied from the EventHubs partition manager. Now that we have
sufficiently deviated from this code, it did not seem worth updating the
abstraction, since we are unlikely to move away from blob leases.