Fix null reference exception in DurableTask.AzureStorage
A null value for OrchestrationInstance on the runtimeState object in
AzureStorageOrchestrationService.CompleteTaskOrchestrationWorkItemAsync
causes rare null reference exceptions for some customers. Unfortunately,
due to the lack of helpful information when this exception is
encountered, we don't know what scenarios cause the orchestration
instance to be null.

This PR currently just adds null coalescion + a warning log. It's
possible that a null execution id that this results in will cause issues
down the line in this method, but this will at least make sure we know
which orchestration instance hit this code path, which makes it far
easier to diagnose the issue.