Fixed a message order regression caused by the de-dupe PR (#541)
The key issue being fixed is a problem where execution started messages are executed out of order after #528 was merged. None of the existing tests were sensitive to this. However, some of the Durable Functions tests were. To fix this, we move the deduplication of execution started messages back onto the main thread, though there may be a potential perf hit.

Customers can turn off this deduplication if performance is too highly impacted.