Change single multi-instance table query to many parallel queries (#568)
In an attempt to reduce the latency + storage transaction impact of the
new deduplication of ExecutionStarted events, we tried to query the
status of multiple instance ids in a single query. Unfortunately, for
certain scenarios, this resulted in very large latency requests.

In order to reduce latency, we will now make an individual request for
each instance id (we do make these requests in parallel though).