Prepare DurableTask.Core and DurableTask.AzureStorage release
The previous fix #560 was not actually released as the fix is contained
in DurableTask.Core, which was not released as a part of
DurableTask.AzureStorage v1.8.8.

The fix in #560 is more severe than we thought as it makes a
TaskHubWorker lose an entire partition when it completes an
orchestration, so we are also releasing a new version of
DurableTask.AzureStorage that depends on DurableTask.Core having this
fix.