Add SBOM to DurableTask.AzureStorage Release NuGet package (#684)
This PR adds the SBOM manifest to the DurableTask.AzureStorage Release NuGet package. It adds a dotnet publish step and moves the manifest generator step to come before the dotnet pack step. It also adds a .nuspec file that specifies the additional files that need to get added to the NuGet package during dotnet pack.