Support for new error propagation mode (#681)
When TaskHubWorker.ErrorPropagationMode property is set to ErrorPropagationMode.UseFailureDetails, instead of serializing and deserializing exceptions for tasks and activities, we'll create a TaskFailureDetails object with the exception type, message, and stack trace. This ensures that orchestrations can know about all types of exceptions and not just serializeable exceptions. The default behavior is unchanged.