Enable Tags to be surfaced via TaskHubClient
This change adds a property to the Azure tracking table ("Tags") that will store the tags for an orchestration. If it is too large, it is offloaded to a blob rather than being stored as a property.

As part of this change, `OrchestrationInstanceStatus` from the table is no longer created with reflection; instead, properties are searched for and then set on the status DTO.

Fixes #840