Expose middleware properties in activity/orchestrator
The properties bag from DispatchMiddlewareContext contains information
that might be useful from activities/orchestrators. However, they
currently do not show up from those context options. This change plumbs
the dictionary through and surfaces it for both activities and
orchestrators.

As part of this, a new interface IContextProperties is added that just
has the properties dictionary. However, this allows for shared code to
set and get properties by type or by name.