Update DistributedTraceContext (#969)
This PR adds a setter for TraceParent to allow it to get serialized. Without this addition, we weren't able to run Netherite apps with the distributed tracing changes because we were seeing this exception: InvalidDataContractException: No set method for property 'TraceParent' in type 'DurableTask.Core.Tracing.DistributedTraceContext'.

It also updates the TraceState setter to include this condition - traceState?.Length <= 513 ? traceState : null instead of adding it in the constructor.