Add versioning to DTFx orchestration dispatch
This commit adds versioning similar to the feature created in
durabletask-dotnet. This portion of versioning allows for an
orchestration to be failed or abandonned based on the configuration
provided.

The actual configuration will be provided in a follow-up PR in the
web extensions as it should be handled through the host.json file.

Signed-off-by: halspang <halspang@microsoft.com>