Fix history events buildup from version reject
When an orchestration was being rejected due to a versioning mismatch
an issue would cause the history events of the orchestration to build
continuously. This was caused by the orchestration adding an
OrchestrationStarted event before the versioning took place and was
saved even on a rejected version.

This change causes the versioning code to exit the process method
entirely instead of just prematurely end the process loop. This avoids
the commit of the history.

Signed-off-by: halspang <halspang@microsoft.com>